Callgrind Integration
A new interface allows the call graph to be built up from a callgrind
output.

Tests have been added and the interface has been added to the existing
profile binary.

The new script "readCallGrind.sh" wraps the awk pre-processor and
calling profile with the correct arguments.

TODO:
  - Better libExec integration
  - Better handling of C++ names (they're too long navigate easily atm)
  - Readline integration (will also help with the c++ names issue)

NOTE: For sensisble results the --separate-callers option should be
      parsed to callgrind, otherwise it does not produce a true call
      graph