Add cgv binary
A new binary, to replace the old profile binary, takes a more
configurable approach to navigating the callgraph.

Using the same core data structures as profile the new Commands library
has been levaraged to allow users to configure their owns commands. This
simplifies the application code and reduces the number of utility
functions that must be defined in the binary .