Integrate gcgv with the build process
The new CEF binary has been integrated with the dev_tools build process.
This is the first stage in integrating the binary with call-grind code.

DEV_TOOLS has been updated to support CEF applications - the only
requriement is that the environment variable CEF_DIR points to the CEF
installation directory (which has include, Release, Debug and Resources
directly inside it).