Initial Node <-> CEF integration
Initial integration between the callgraph data structure and the CEF
GUI is now provided.

The Node tree is held in the entirely browser process, and queries are
provided on the request-reply interface to update and interogate that
data. This allows the GUI to remain responsive whilst processing data,
taking advantage of the in-built process split rather than forcing us to
manage a new thread.

It is now possible, upon initialisation, to:
  - Get a list of files matching a glob pattern (used to drive a file
    picker box on the loading page)
  - Attempt to load a callgrind output file, returning an error if it is
    invalid.

Once a valid callgraph has been loaded it is now possible to:
  - Interogate the details of the file (name, units captured)
  - Get the flat view from the current working directory (which may be
    filtered by depth and / or a regex string)

The new test-app interface from DEV_TOOLS has been used to create a
build-time test which validates this java-script integration is working
as expected. A minor switch was added to the client to prevent it trying
to install GTK window handling and browser management into the test app.