Initial working version of the GCGV GUI.
gcgv has now been updated to leverage the javascript interface allowing
it to load and use call-graph files. The loading screen is now fully
functional, with a search box to find the file, which can the be used to
start the app proper if it is a valid callgrind file.

(Providing a full file dialogue requires GTK hackery, and is left as a
future enhancement).

The initial analyse screen is provided as a proof of concept. It is
possible to view the table flat-view from the root node, configuring the
depth it is calculated to, and a regex filter to apply.