Navigation enhancements to the GUI
It is now possible to navigate through the call-graph from the CEF GUI.
In particular:
  - Selecting a row in the table view jumps to that node, unless there
    are multiple nodes it could jump to, in which case it opens the
    finder
  - The finder is correctly populated and can be used to navigate to the
    correct ndoe when there are multiple options.
  - The PWD breadcrumb is populated allowing the user to go back up
    towards the ROOT node.

To Fasciliate the GUI changes, new requests have been added which:
   - Search for child nodes, up to a specified depth.
   - Change the current node based on a path.
In addition the details request has been enhanced to supply the current
working directory.

Other than an unhandled exception bug fix (pulled in by the updated to
DEV_TOOLS) there was no changed required to the underlying CEF
integration library.