Migrate from libCef to libWebSockets
The GCGV application has been migrated away from libCEF to use
libWebSockets. (See the DEV_TOOLS commit for further details on the
reasons for making this change).

To do the migration the following changes were made:
   - Removal of CEF machinery and bookkeeping code.
   - Migrated Request handlers to use the new interface provided by
     libWebSockets
   - Updated all gcgv tests to directly test the response JSON using
     libJSON. (Previously this was done by executing Javascript in a
     test CEF application).

As a result of this change, the gcgv binary will no longer create a
window for the user to navigate with. Instead it will listen on a port
for the user to start the static web-application manually from a
web browser. (This process will be eventually scripted, abstracting this
from the user).